phase modified to be complete

+++ server/concepts/phase.ts
@@ -1,28 +1,34 @@
 import { ObjectId } from "mongodb";
 
 import DocCollection, { BaseDoc } from "../framework/doc";
-import { NotAllowedError, NotFoundError } from "./errors";
+import { BadValuesError, NotAllowedError, NotFoundError } from "./errors";
 
-export interface PhaseDoc extends BaseDoc {
-  key: ObjectId;
+export interface PhaseDoc extends NewPhaseDoc {
   deadline: Date;
 }
 
+export interface NewPhaseDoc extends BaseDoc {
+  key: ObjectId;
+  curPhase: number;
+}
+
 export default class PhaseConcept {
+  private readonly store = new DocCollection<NewPhaseDoc>("store phases");
   public readonly active = new DocCollection<PhaseDoc>("active phases");
   public readonly expired = new DocCollection<PhaseDoc>("expired phases");
+  private maxPhase = 3;
+  private deadlineExtension = 24;
 
   /**
-   * Creates a new active phase for a given item that expires at a given time
+   * Creates a new store phase for a given item
    * @param key id of the object
-   * @param deadline date and time when the phase should expire
    * @returns an object containing a success message and the phase object
    */
-  async initialize(key: ObjectId, deadline: Date) {
+  async initialize(key: ObjectId) {
+    await this.expireOld();
     await this.alreadyExists(key);
-    await this.alreadyExpired(deadline);
-    const _id = await this.active.createOne({ key, deadline });
-    return { msg: "Active phase successfully created!", phase: await this.active.readOne({ _id }) };
+    const _id = await this.active.createOne({ key, curPhase: 0 });
+    return { msg: "New stored phase successfully created!", phase: await this.store.readOne({ _id }) };
   }
 
   /**
@@ -45,11 +51,25 @@ export default class PhaseConcept {
   }
 
   /**
+   * Edits the deadline for a given item if the new deadline
+   * hasn't passed yet
+   * @param key id of the item
+   * @param newDeadline new date (should be in the future)
+   * @returns an object containing a success message
+   */
+  async editDeadline(key: ObjectId, newDeadline: Date) {
+    this.alreadyExpired(newDeadline);
+    await this.active.updateOne({ key }, { deadline: newDeadline });
+    return { msg: "Successfully updated the deadline" };
+  }
+
+  /**
    * Removes the active phase by key (could be =0 or =1)
    * @param key id of the item being deleted
    * @returns an object containing a success message
    */
   async deleteActive(key: ObjectId) {
+    await this.expireOld();
     await this.active.deleteOne({ key });
     return { msg: "Active phase deleted successfully!" };
   }
@@ -65,16 +85,79 @@ export default class PhaseConcept {
   }
 
   /**
+   * Change the max phase
+   * @param newMax a positive integer
+   * @throws BadValuesError if newMax is 0 or less or isn't an integer
+   */
+  changeMaxPhase(newMax: number) {
+    if (newMax > 0 && Number.isInteger(newMax)) {
+      this.maxPhase = newMax;
+      return { msg: "Successfully updated max phase value" };
+    }
+    throw new BadValuesError(newMax + " must be an integer greater than 0");
+  }
+
+  /**
+   * Change the deadline extending (hours)
+   * @param newVal number of hours that the deadline will be
+   * extended by
+   * @throws BadValuesError if newVal is 0 or less
+   */
+  changeDeadlineExtension(newVal: number) {
+    if (newVal > 0) {
+      this.deadlineExtension = newVal;
+      return { msg: "Successfully updated deadline extension value" };
+    }
+    throw new BadValuesError(newVal + " must be greater than 0");
+  }
+
+  /**
+   * Adds new phase 1 items from the store into the active group
+   */
+  private async start() {
+    const existingPhase1s = await this.active.readMany({ curPhase: 1 });
+    if (existingPhase1s.length < 2) {
+      const deadline = new Date();
+      deadline.setTime(deadline.getTime() + this.deadlineExtension * 60 * 60 * 1000);
+
+      const phase1 = await this.store.readMany({}, { limit: 2 });
+      for (const item of phase1) {
+        await this.active.createOne({ key: item.key, deadline, curPhase: 1 });
+        await this.store.deleteOne({ _id: item._id });
+      }
+    }
+  }
+
+  /**
    * Moves all expired phases that are currently in active into expired
+   * if maxed out on the phase or progresses the phase and deadline
    */
   private async expireOld() {
     const now = new Date();
     const expired = await this.active.readMany({ deadline: { $lt: now } });
 
     for (const phase of expired) {
-      await this.expired.createOne({ key: phase.key, deadline: phase.deadline });
-      await this.active.deleteOne({ _id: phase._id });
+      if (phase.curPhase === this.maxPhase) {
+        await this.expired.createOne({ key: phase.key, deadline: phase.deadline });
+        await this.active.deleteOne({ _id: phase._id });
+      } else {
+        const newDate: Date = phase.deadline;
+        let newPhase = phase.curPhase;
+
+        while (newDate < now && newPhase < this.maxPhase) {
+          newDate.setTime(newDate.getTime() + this.deadlineExtension * 60 * 60 * 1000);
+          newPhase += 1;
+        }
+
+        if (newDate < now) {
+          await this.expired.createOne({ key: phase.key, deadline: phase.deadline });
+          await this.active.deleteOne({ _id: phase._id });
+        } else {
+          await this.active.updateOne({ _id: phase._id }, { deadline: newDate, curPhase: newPhase });
+        }
+      }
     }
+    await this.start();
   }
 
   /**
@@ -96,7 +179,8 @@ export default class PhaseConcept {
    */
   private async alreadyExists(key: ObjectId) {
     const phase = await this.active.readOne({ key });
-    if (phase) {
+    const phaseStored = await this.store.readOne({ key });
+    if (phase || phaseStored) {
       throw new KeyExistsError(key);
     }
   }
@@ -117,7 +201,7 @@ export default class PhaseConcept {
 
 export class KeyExistsError extends NotAllowedError {
   constructor(public readonly key: ObjectId) {
-    super("{0} already has an active phase!", key);
+    super("{0} already has an active or stored phase!", key);
   }
 }
